+++ Src/FileUtility.h
@@ -6,7 +6,7 @@ namespace Catherine
 	class FileUtility
 	{
 	public:
-		static unsigned int LoadFileContent(const char * param_Path, void ** param_Ptr);
+		static size_t LoadFileContent(const char * param_Path, void ** param_Ptr);
 		static std::string GetDictionary(const std::string & path);
 	};
 }
